--- a/change/@fluentui-api-docs-44f3b6cf-e2ae-452b-bf50-ef1af35445bb.json
+++ b/change/@fluentui-api-docs-44f3b6cf-e2ae-452b-bf50-ef1af35445bb.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "chore: bump api-extractor-model in order to migrate to TS 4.9",
+  "packageName": "@fluentui/api-docs",
+  "email": "martinhochel@microsoft.com",
+  "dependentChangeType": "patch"
+}