--- a/packages/fluentui/react-northstar/src/components/Tooltip/TooltipContent.tsx
+++ b/packages/fluentui/react-northstar/src/components/Tooltip/TooltipContent.tsx
@@ -9,7 +9,7 @@ import {
   ForwardRefWithAs,
 } from '@fluentui/react-bindings';
 import * as customPropTypes from '@fluentui/react-proptypes';
-import * as PopperJs from '@popperjs/core';
+import type { BasePlacement as PopperJsBasePlacement, Placement as PopperJsPlacement } from '@popperjs/core';
 import * as PropTypes from 'prop-types';
 import * as React from 'react';
 
@@ -49,7 +49,7 @@ export interface TooltipContentProps extends UIComponentProps, ChildrenComponent
 }
 
 export type TooltipContentStylesProps = Required<Pick<TooltipContentProps, 'pointing' | 'open' | 'subtle'>> & {
-  basePlacement: PopperJs.BasePlacement;
+  basePlacement: PopperJsBasePlacement;
 };
 
 export const tooltipContentClassName = 'ui-tooltip__content';
@@ -127,7 +127,7 @@ TooltipContent.displayName = 'TooltipContent';
 
 TooltipContent.propTypes = {
   ...commonPropTypes.createCommon(),
-  placement: PropTypes.oneOf<PopperJs.Placement>([
+  placement: PropTypes.oneOf<PopperJsPlacement>([
     'auto-start',
     'auto',
     'auto-end',