--- a/packages/fluentui/react-northstar/src/utils/positioner/getBasePlacement.ts
+++ b/packages/fluentui/react-northstar/src/utils/positioner/getBasePlacement.ts
@@ -1,12 +1,12 @@
-import * as PopperJs from '@popperjs/core';
+import type { BasePlacement as PopperJsBasePlacement, Placement as PopperJsPlacement } from '@popperjs/core';
 
-const rtlMapping: Partial<Record<PopperJs.BasePlacement, PopperJs.BasePlacement>> = {
+const rtlMapping: Partial<Record<PopperJsBasePlacement, PopperJsBasePlacement>> = {
   left: 'right',
   right: 'left',
 };
 
-export function getBasePlacement(placement: PopperJs.Placement | undefined, rtl: boolean): PopperJs.BasePlacement {
-  const basePlacement = (placement || '').split('-', 1).pop() as PopperJs.BasePlacement;
+export function getBasePlacement(placement: PopperJsPlacement | undefined, rtl: boolean): PopperJsBasePlacement {
+  const basePlacement = (placement || '').split('-', 1).pop() as PopperJsBasePlacement;
 
   return (rtl && rtlMapping[basePlacement]) || basePlacement;
 }