--- a/packages/fluentui/react-northstar/src/utils/positioner/getBoundary.ts
+++ b/packages/fluentui/react-northstar/src/utils/positioner/getBoundary.ts
@@ -1,4 +1,4 @@
-import * as PopperJs from '@popperjs/core';
+import type { Boundary as PopperJsBoundary } from '@popperjs/core';
 
 import { getScrollParent } from './getScrollParent';
 import { Boundary } from './types';
@@ -6,7 +6,7 @@ import { Boundary } from './types';
 /**
  * Allows to mimic a behavior from V1 of Popper and accept `window` and `scrollParent` as strings.
  */
-export function getBoundary(element: HTMLElement, boundary: Boundary): PopperJs.Boundary {
+export function getBoundary(element: HTMLElement, boundary: Boundary): PopperJsBoundary {
   if (boundary === 'window') {
     return element.ownerDocument.documentElement;
   }