--- a/packages/fluentui/react-northstar/src/utils/positioner/isIntersectingModifier.ts
+++ b/packages/fluentui/react-northstar/src/utils/positioner/isIntersectingModifier.ts
@@ -1,4 +1,5 @@
-import { detectOverflow, Modifier } from '@popperjs/core';
+import { detectOverflow } from '@popperjs/core';
+import type { Modifier } from '@popperjs/core';
 
 export const isIntersectingModifier: IsIntersectingModifier = {
   name: 'is-intersecting-modifier',