--- a/packages/fluentui/react-northstar/src/utils/positioner/partitionPopperPropsFromShorthand.ts
+++ b/packages/fluentui/react-northstar/src/utils/positioner/partitionPopperPropsFromShorthand.ts
@@ -1,7 +1,7 @@
 import { ObjectShorthandValue, ShorthandValue } from '../../types';
 import { PopperShorthandProps } from './types';
 
-export function partitionPopperPropsFromShorthand<P>(
+export function partitionPopperPropsFromShorthand<P extends {}>(
   value: ShorthandValue<P & { popper?: PopperShorthandProps }>,
 ): [ShorthandValue<P> | ObjectShorthandValue<P>, PopperShorthandProps | undefined] {
   if (typeof value === 'object' && value !== null && !Array.isArray(value)) {