<!--
Thank you for submitting a pull request!

Please verify that:
* [ ] Code is up-to-date with the `master` branch
* [ ] Your changes are covered by tests (if possible)
* [ ] You've run `yarn change` locally


PR flow tips:
* [ ] Try to start with a Draft PR
* [ ] Once you're ready (ideally the pipeline is passing) promote your PR to Ready for Review. This step will auto-assign reviewers for your PR.
-->

## Previous Behavior

<!-- This is the behavior we have today -->

## New Behavior

- 4.8 https://www.typescriptlang.org/docs/handbook/release-notes/typescript-4-8.html
- 4.9 https://www.typescriptlang.org/docs/handbook/release-notes/typescript-4-9.html 

**api-extractor:**

- bumps to last api-extractor version that support TS 4.8 -> we are setting yarn resolutions to override this to use monorepo TS (4.9.5)
- api-extractor-model and  ts-doc bumps to match api-extractor version for deduplication

**web-components**

- html-react-parser used within vr tests ships invalid types for `nodeResolution: node16` setup. Simplest solution was to provide local type override for `parse` api usage
  - more context https://github.com/remarkablemark/html-react-parser/issues/1305#issuecomment-2136962278 

**react-northstar**

- as v9 package (we use `strict` type checking in all v9) relies on v0 ( `react-migration-v0-v9` ), type errors were exposed when generating api.md:
  - unconstrained generics issues https://www.typescriptlang.org/docs/handbook/release-notes/typescript-4-8.html#unconstrained-generics-no-longer-assignable-to- 
  - also `@popperjs/core` has not ts 4.9 compliant types and only 3 recent versions mitigate this issue - that's why I needed to bump it ( which increases v0 bundle size )
  - there is a 1% bundle size increase in some packages - seems related to poppejs which size increased by 3kB ( from 17.6 kB -> 20.9 kB )


### Perf bench

> `yarn lage type-check`

| Before (4.7) | After (4.9) |
|--------|--------|
| 12m 11s | 11m 7s / **𝚫 9% 🚅 faster** | 

## Related Issue(s)

- https://github.com/remarkablemark/html-react-parser/issues/1305

<!-- Please link the issue being fixed so it gets closed when this is merged. -->

- Fixes #